Fix cleanup_images

+++ cleanup_images.rb
@@ -1,4 +1,4 @@
 #!/usr/bin/ruby
 
 imagesDbs = Dir.glob('./clusters/**/*.db').each { |imagesDb| File.unlink(imagesDb) }
-File.unlink('./templates.db')
+File.unlink('./templates.db') if File.exist? './templates.db'
